bedwars profile type & comments on config

+++ celestialspigot-server/src/main/java/net/minecraft/server/EntityProjectile.java
@@ -141,7 +141,7 @@ public abstract class EntityProjectile extends Entity implements IProjectile {
 
                 if (entity1.ad() && (entity1 != entityliving || this.ar >= 5)) {
                     float f = 0.3F;
-                    AxisAlignedBB axisalignedbb = entity1.getBoundingBox().grow((double) f, (double) f, (double) f);
+                    AxisAlignedBB axisalignedbb = entity1.getBoundingBox().grow(f, f, f);
                     MovingObjectPosition movingobjectposition1 = axisalignedbb.a(vec3d, vec3d1);
                     if (healPotion && CelestialSpigot.INSTANCE.getConfig().isSmoothHealPotions() && movingobjectposition1 == null && entity1 == entityliving && ticksLived % 3 == 0 && !entity1.inWater && !entity1.ab() && !entity1.isSneaking()) {
                         movingobjectposition1 = new MovingObjectPosition(entity1);
